Extracted prototypes to their own modules in pageobject subpackage
pageobject.prototypes now imports classes from the new prototypes modules, so syntax remains the same
